% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Standard Deviation
This function calculates the standard deviation of input vector.}
\usage{
standard_deviation(x)
}
\arguments{
\item{vector}{x}
}
\value{
numeric
}
\description{
Standard Deviation
This function calculates the standard deviation of input vector.
}
\examples{
standard_deviation(c(2,6,7,8))
}

