% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_error.R
\name{standard_error}
\alias{standard_error}
\title{Standard Error
This function calculates the standard error of input vector}
\usage{
standard_error(x)
}
\arguments{
\item{vector}{x}
}
\value{
numeric
}
\description{
Standard Error
This function calculates the standard error of input vector
}
\examples{
standard_error(c(1,2,3,4))
}

